Highlighting lexer - replaced long with int for remaining 'data' parameters and local declarations

+++ platform/core-api/src/com/intellij/lexer/RestartableLexer.java
@@ -37,9 +37,9 @@ public interface RestartableLexer {
 
   IElementType unpackTokenData(int data);
 
-  int unpackState(long data);
+  int unpackState(int data);
 
-  boolean containsRestartableState(long data);
+  boolean containsRestartableState(int data);
 
   boolean isRestartableState(int state);
 }
